refactored the code, following the separation of concern principles, with the intent to make the code easy to follow
